+++ openccu_loom_client/model/__init__.py
@@ -17,6 +17,7 @@ through.
 
 from __future__ import annotations
 
+from openccu_loom_client.model.alarm_panel import MASTER_AREA_ID, AlarmPanel
 from openccu_loom_client.model.channel import Channel
 from openccu_loom_client.model.custom_data_point import CustomDataPoint
 from openccu_loom_client.model.data_point import DataPoint
@@ -26,10 +27,12 @@ from openccu_loom_client.model.sysvar import Sysvar
 
 __all__ = [
     # General
+    "AlarmPanel",
     "Channel",
     "CustomDataPoint",
     "DataPoint",
     "Device",
+    "MASTER_AREA_ID",
     "Program",
     "Sysvar",
 ]